requirements.txt
discord.py>=1.7.3
jishaku>=2.0.0
